feat: add user and trip existence validation to TripService methods

+++ src/trip/application/trip.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { TripRepository } from '../domain/repositories/trip.repository';
 import {
   CreateTripServiceInputInterface,
@@ -7,14 +7,24 @@ import {
   ListTripsByUserIdServiceInputInterface,
   ListTripsByUserIdServiceOutputInterface,
 } from './trip.service.interface';
+import { UserRepository } from 'src/shared/repositories/user.repository';
 
 @Injectable()
 export class TripService {
-  constructor(private readonly tripRepository: TripRepository) {}
+  constructor(
+    private readonly tripRepository: TripRepository,
+    private userRepository: UserRepository,
+  ) {}
 
   async createTripService(
     data: CreateTripServiceInputInterface,
   ): Promise<CreateTripServiceOutputInterface> {
+    const user = await this.userRepository.findById(data.userTrips.userId);
+
+    if (!user) {
+      throw new NotFoundException(`User not found.`);
+    }
+
     const formattedDestinations = data.destination.map((dest) => ({
       ...dest,
       startDate: new Date(dest.startDate),
@@ -35,6 +45,12 @@ export class TripService {
     const page = listTripsByUserId.page ? Number(listTripsByUserId.page) : 1;
     const limit = listTripsByUserId.limit ? Number(listTripsByUserId.limit) : 9;
     const skip = (page - 1) * limit;
+
+    const user = await this.userRepository.findById(listTripsByUserId.userId);
+
+    if (!user) {
+      throw new NotFoundException(`User not found.`);
+    }
 
     const trips = await this.tripRepository.findTripsByUserId({
       userId: listTripsByUserId.userId,
@@ -67,12 +83,24 @@ export class TripService {
   }
 
   async deleteTrip(tripId: string) {
+    const trip = await this.tripRepository.findTripById(tripId);
+
+    if (!trip) {
+      throw new NotFoundException(`Trip not found`);
+    }
+
     await this.tripRepository.deleteTripByTripId(tripId);
   }
 
   async getTripByIdService(
     tripId: string,
   ): Promise<GetTripByUserIdServiceOutputInterface> {
+    const uniqueTrip = await this.tripRepository.findTripById(tripId);
+
+    if (!uniqueTrip) {
+      throw new NotFoundException(`Trip not found`);
+    }
+
     const trip = await this.tripRepository.getTripById(tripId);
 
     return trip;
